Switch from asyncio to regular blocking functions
Asyncio is great and Azure Function's own support for it works well,
however this limits us in the number of third-party libraries we can
use.

Sentry's lack of support for it is what gave me the final push to
rewrite this back to non-async code, primarily because this isn't going
to be a high-preformance/high utilization project anyway so threads and
blocking is fine.